% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Apply a power}
\usage{
pow(x, p)
}
\arguments{
\item{x}{Vector, numeric, to apply power to}

\item{p}{Exponent}
}
\value{
The vector should be returned in the form/function
}
\description{
Meant for internal use only
}
