% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powers-package.R
\docType{package}
\name{Powers-package}
\alias{Powers}
\alias{Powers-package}
\title{Powers: Apply power to vector}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Detais go here for future documentation about the package
}
\keyword{internal}
